+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 - Introduced `update(ByteBuffer)`.
+- Introduced `done(out: OutputStream, len: Int)`.
 
 ## [1.9.0] - 2020-11-21
 - Significant performance improvement and reducing memory footprint at `done`.